requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.0
Flask-Limiter==3.8.0
Flask-Swagger-UI==4.11.1
gunicorn==22.0.0
psycopg[binary,pool]==3.2.3
requests==2.32.3
python-dotenv==1.0.1
